+++ store/iavl/wire.go
@@ -1,7 +1,0 @@
-package iavl
-
-import (
-	"github.com/cosmos/cosmos-sdk/codec"
-)
-
-var cdc = codec.New()