+++ x/auth/signing/verify_test.go
@@ -49,8 +49,7 @@ func TestVerifySignature(t *testing.T) {
 		AccountNumber:   acc.GetAccountNumber(),
 		AccountSequence: acc.GetSequence(),
 	}
-	signBytes := types.StdSignBytes(signerData.ChainID, signerData.AccountNumber, signerData.AccountSequence,
-		fee, msgs, memo)
+	signBytes := types.StdSignBytes(signerData.ChainID, signerData.AccountNumber, signerData.AccountSequence, 10, fee, msgs, memo)
 	signature, err := priv.Sign(signBytes)
 	require.NoError(t, err)
 
@@ -60,6 +59,7 @@ func TestVerifySignature(t *testing.T) {
 
 	handler := MakeTestHandlerMap()
 	stdTx := types.NewStdTx(msgs, fee, []types.StdSignature{stdSig}, memo)
+	stdTx.TimeoutHeight = 10
 	err = signing.VerifySignature(pubKey, signerData, sigV2.Data, handler, stdTx)
 	require.NoError(t, err)
 
@@ -67,8 +67,7 @@ func TestVerifySignature(t *testing.T) {
 	multisigKey := multisig.NewPubKeyMultisigThreshold(2, pkSet)
 	multisignature := multisig.NewMultisig(2)
 	msgs = []sdk.Msg{testdata.NewTestMsg(addr, addr1)}
-	multiSignBytes := types.StdSignBytes(signerData.ChainID, signerData.AccountNumber, signerData.AccountSequence,
-		fee, msgs, memo)
+	multiSignBytes := types.StdSignBytes(signerData.ChainID, signerData.AccountNumber, signerData.AccountSequence, 10, fee, msgs, memo)
 
 	sig1, err := priv.Sign(multiSignBytes)
 	require.NoError(t, err)