+++ x/staking/keeper/common_test.go
@@ -18,7 +18,7 @@ var (
 
 // createTestInput Returns a simapp with custom StakingKeeper
 // to avoid messing with the hooks.
-func createTestInput() (*codec.Codec, *simapp.SimApp, sdk.Context) {
+func createTestInput() (*codec.LegacyAmino, *simapp.SimApp, sdk.Context) {
 	app := simapp.Setup(false)
 	ctx := app.BaseApp.NewContext(false, tmproto.Header{})